refactor SimSearchService, VectorIndexService, EmbeddingService
