+++ .vscode/launch.json
@@ -71,12 +71,6 @@
       "console": "integratedTerminal",
       "cwd": "${workspaceFolder}/backend/src",
       "justMyCode": true,
-      "env": {
-        "PYTHONPATH": "${workspaceFolder}/backend/src",
-        "HUGGINGFACE_HUB_CACHE": "${workspaceFolder}/docker/models_cache",
-        "HF_HOME": "${workspaceFolder}/docker/models_cache",
-        "TORCH_HOME": "${workspaceFolder}/docker/models_cache"
-      },
       "envFile": "${workspaceFolder}/backend/.env"
     },
     {
