+++ backend/src/app/core/analysis/cota/pipeline/steps/finetune_apply_compute.py
@@ -2,12 +2,9 @@ from typing import Dict, List, Tuple
 
 import numpy as np
 from app.core.analysis.cota.pipeline.cargo import Cargo
-from app.core.data.dto.concept_over_time_analysis import (
-    COTAConcept,
-    COTASentence,
-)
+from app.core.data.dto.concept_over_time_analysis import COTAConcept, COTASentence
 from app.core.data.repo.repo_service import RepoService
-from app.core.db.simsearch_service import SimSearchService
+from app.core.db.vector_index_service import VectorIndexService
 from app.preprocessing.ray_model_service import RayModelService
 from app.preprocessing.ray_model_worker.dto.cota import (
     RayCOTAJobInput,
@@ -19,7 +16,7 @@ from umap.umap_ import UMAP
 
 rms: RayModelService = RayModelService()
 repo: RepoService = RepoService()
-sims: SimSearchService = SimSearchService()
+index: VectorIndexService = VectorIndexService()
 
 
 def finetune_apply_compute(cargo: Cargo) -> Cargo:
@@ -45,7 +42,7 @@ def finetune_apply_compute(cargo: Cargo) -> Cargo:
                 sentence.concept_similarities[concept_id] = similarity
 
     else:
-        embeddings_tensor = sims.get_sentence_embeddings(
+        embeddings_tensor = index.get_embeddings(
             search_tuples=[
                 (cota_sent.sentence_id, cota_sent.sdoc_id) for cota_sent in search_space
             ]
