+++ backend/src/app/core/db/vector_index_service.py
@@ -2,6 +2,8 @@ from abc import ABC, abstractmethod
 from typing import Iterable, List, Sequence, Tuple, Union, overload
 
 import numpy as np
+from config import conf
+from loguru import logger
 
 from app.core.data.dto.search import SimSearchDocumentHit, SimSearchSentenceHit
 from app.core.db.index_type import IndexType
@@ -9,6 +11,34 @@ from app.util.singleton_meta import SingletonMeta
 
 
 class VectorIndexService(ABC, metaclass=SingletonMeta):
+    def __new__(cls, reset_vector_index=False):
+        index_name: str = conf.vector_index.service
+        match index_name:
+            case "qdrant":
+                # import and init QdrantService
+                from app.core.db.qdrant_service import QdrantService
+
+                return QdrantService(flush=reset_vector_index)
+            case "typesense":
+                # import and init TypesenseService
+                from app.core.db.typesense_service import TypesenseService
+
+                return TypesenseService(flush=reset_vector_index)
+            case "weaviate":
+                # import and init WeaviateService
+                from app.core.db.weaviate_service import WeaviateService
+
+                instance = super(VectorIndexService, WeaviateService).__new__(
+                    WeaviateService
+                )
+                return instance
+            case _:
+                msg = (
+                    f"VECTOR_INDEX environment variable not correctly set: {index_name}"
+                )
+                logger.error(msg)
+                raise SystemExit(msg)
+
     @abstractmethod
     def add_embeddings_to_index(
         self,
@@ -84,9 +114,10 @@ class VectorIndexService(ABC, metaclass=SingletonMeta):
         pass
 
     @abstractmethod
-    def get_sentence_embeddings(
+    def get_embeddings(
         self,
         search_tuples: List[Tuple[int, int]],
+        index_type: IndexType = IndexType.SENTENCE,
     ) -> np.ndarray:
         pass
